[pypmc] Move ``_partition.py`` from ``tools`` to ``density`` to avoid circular dependence
