+++ pypmc/density/__init__.py
@@ -4,3 +4,4 @@
 
 # import these submodules by default
 from . import gauss, student_t, mixture, base
+from ._partition import partition, patch_data
