+++ pypmc/density/partition_test.py
